Start work on expanding edges & new vertices
The opposite of the rejected artist animations, I'm now working on
the expanding edges and fading-in vertices when related artists are
populated.

This bit is tricky because of the fact that we don't know what the
state will become when the animation starts (that info comes from
spotify), and it's still pretty buggy.